added check for #destroy if like exists

+++ app/controllers/posts/likes_controller.rb
@@ -7,8 +7,8 @@ class Posts::LikesController < Posts::ApplicationController
   end
 
   def destroy
-    like = resource_post.likes.find_by!(user: current_user)
-    like.destroy!
+    like = resource_post.likes.find_by(user: current_user)
+    like.destroy! if like
     # [TODO] message?..
 
     redirect_to resource_post
